+++ dali/operators/util/property.cc
@@ -42,7 +42,7 @@ template <>
 void SourceInfo<GPUBackend>::FillOutput(workspace_t<GPUBackend>& ws) {
   const auto& input = ws.template Input<GPUBackend>(0);
   auto& output = ws.template Output<GPUBackend>(0);
-  for (size_t sample_id = 0; sample_id < input.num_samples(); sample_id++) {
+  for (int sample_id = 0; sample_id < input.num_samples(); sample_id++) {
     auto si = GetSourceInfo(input, sample_id);
     auto output_ptr = output.raw_mutable_tensor(static_cast<int>(sample_id));
     cudaMemcpyAsync(output_ptr, si.c_str(), si.length(), cudaMemcpyDefault, ws.stream());
@@ -53,7 +53,7 @@ template <>
 void Layout<GPUBackend>::FillOutput(workspace_t<GPUBackend>& ws) {
   const auto& input = ws.template Input<GPUBackend>(0);
   auto& output = ws.template Output<GPUBackend>(0);
-  for (size_t sample_id = 0; sample_id < input.num_samples(); sample_id++) {
+  for (int sample_id = 0; sample_id < input.num_samples(); sample_id++) {
     auto layout = GetLayout(input, sample_id);
     auto output_ptr = output.raw_mutable_tensor(static_cast<int>(sample_id));
     cudaMemcpyAsync(output_ptr, layout.c_str(), layout.size(), cudaMemcpyDefault, ws.stream());
